Display logs for unattended installs
When a program is taking a long time to finish, there are two
possibilities:

1. The program got stuck and will never complete.
2. The program is just taking a long time and will finish eventually.

It’s impossible to know for sure which one the program is doing, but the
more information you have, the easier it is to make an educated guess.
The goal of this commit is to give the user more information about
unattended-install.service so that they can make an educated guess about
whether or not it’s working. Additionally, this commit makes it easier
for users to see how close the installation is to finishing.