Make installation ISOs use a newer kernel
The main motivation behind this commit is to make it easier to make a
future commit. That future commit will switch one of the systems to
bcachefs [1]. At the moment, pkgs.linux gives you Linux 6.1.78 [2].
Bcachefs didn’t make it into the kernel until Linux 6.7 [3], so I need
to make the installation ISOs use a newer version of Linux before I can
switch to bcachefs.

I chose to use a module from Nixpkgs that also disables ZFS [4] support.
I did this because it was easy (the “-no-zfs” module was right next to
the regular module) and because I’m not going to be using ZFS anyway.
Why waste space on a feature that won’t ever be used?

[1]: <https://bcachefs.org>
[2]: <https://web.archive.org/web/20240218124759/https://search.nixos.org/packages?channel=23.11&show=linux&from=0&size=50&sort=relevance&type=packages&query=linux>
[3]: <https://lore.kernel.org/lkml/CAHk-=widprp4XoHUcsDe7e16YZjLYJWra-dK0hE1MnfPMf6C3Q@mail.gmail.com/>
[4]: <https://openzfs.org/wiki/Main_Page>