Switch jason-desktop-linux to bcachefs
I kept running out of disk space on jason-desktop-linux, so I ordered
new drives. I’m using this as an opportunity to try out a new
filesystem. Specifically, I chose bcachefs because of the way that it
does caching. I don’t have to worry about which files should be put on
HDDs and which files should be put on SSDs. Instead, I can just let
bcachefs keep the most often used files on the SSD and move everything
else to the HDDs.