Remove CDEmu
I had originally enabled CDEmu to allow me to run a program in Wine [1].
It’s been a long time since I’ve run anything in Wine, much less a
program that requires a CD.

The main motivation behind this commit is to make it easier to create a
future commit. At the moment, I’m not able to use the latest version of
Linux because one of Nixpkgs’s CDEmu-related packages isn’t compatible
with the latest version of Linux [2]. Removing CDEmu now will make it
possible for that future commit to switch me over to the latest version
of Linux.

[1]: 5bf2a42 (Enable CDemu, 2023-04-06)
[2]: src/modules/bcachefs.nix