Replace install ISOs with install drives
The idea behind d8fb11c (Include prebuilt config on install ISOs,
2024-01-31) was simple: save time by not downloading things that have
already been downloaded. If I always build installation ISOs on
Jason-Desktop-Linux, then I won’t have to keep redownloading things that
are already in Jason-Desktop-Linux’s Nix store. Sure, I might have to
build a few packages here or there, but the majority of the ISO should
already be in the Nix store, so it should save time, right?

As it turns out, I probably wasn’t saving time by doing this. Here’s how
things would have worked if I wasn’t building custom ISOs that contained
prebuilt configs:

1. Jason-Desktop-Linux downloads the minimal install ISO. This causes
about 1GiB of data to be downloaded into RAM and then written to disks.

2. Jason-Desktop-Linux copies the ISO onto a USB drive. In the worst
case scenario, this causes about 1GiB of data to be read into RAM and
then written to the USB drive. In reality, a lot of the data will likely
still be cached, so it won’t have to be reread.

3. The target machine boots off of the USB drive. In the worst case
scenario, this will causes 1GiB of data to be read into RAM and
decompressed [1].

4. The target machine builds and installs the configuration. In the
worst case scenario, this will cause about 100GiB* to be downloaded and
then written to a disk.

In total, that process results in about 104GiB being transferred into
RAM and then written to disks. Now, let’s take a look at how things were
working when I was building custom ISOs that contained prebuilt configs:

1. Jason-Desktop-Linux makes sure that it has a copy of all of the
packages needed on the target system. In the best case scenario, no data
is downloaded.

2. Jason-Desktop-Linux builds a SquashFS image [2]. In the best case
scenario, the SquashFS image is already in the Nix store so no data
needs to be read. In reality, there’s always at least one bit that has
been changed since the last time I created an ISO. As long as there’s at
least one bit that has been changed, every single store path needed by
the target system  has to be read into RAM and then written back to the
disk. In other words, most of the time, about 100GiB has to be read into
RAM and then written back to the disks.

3. If a new SquashFS image had to be generated, then a new ISO image has
to be generated. ISO images contain the SquashFS image [3]. In other
words, most of the time, about 100GiB has to be read back into RAM and
then written back to the disks.

4. Jason-Desktop-Linux copies the ISO onto a USB drive. This results in
about 100GiB being read back into RAM and then written to the USB drive.

5. The target machine boots off of the USB drive. In the worst case
scenario, this will probably cause about 1GiB** of data to be read into
RAM. It doesn’t get decompressed because I disabled compression [4].

6. The target machine builds and installs the configuration. In the
worst case scenario, this will cause about 100GiB* to be downloaded and
then written to the disk.

All of that means that in a typical scenario, we’ll have to wait for
over 300GiB to be transferred into RAM and then written to a disk. In
other words, the typical scenario for building ISOs is probably much
slower than the worst case scenario for just downloading everything.

                                 ✱ ✱ ✱

The fundamental problem here is that we’re generating an image that
later gets written to the USB drive. In an ideal scenario, we would just
copy the data directly to the USB drive. That’s exactly what this commit
does. This commit replaces install ISOs with install drives. Creating an
install drive will cause the data to get written directly to the USB
drive.

Additionally, ISOs are read-only filesystems. This means that if you
want to update an ISO, you have to generate a new one. Install drives
contain filesystems that can be written to. This means that in the
future I’ll be able to update install drives without having to recreate
them.

                                 ✱ ✱ ✱

This commit also contains some tangentially-related changes.
Specifically, I had to rewrite the installation instructions to make
them talk about install drives instead of installation ISOs. While
rewriting them, I decided to move each major step into its own file.
I’ve found that doing so has made following the instructions a lot less
daunting. Also, I renamed src/misc/disko/vm-base.nix to
src/misc/disko/base-disk-configuration.nix because that file is no
longer only being used by VMs.

                                 ✱ ✱ ✱

*I haven’t actually tested this, and I have reason to believe that in
most situations, it will download a lot less. That being said, before
this commit, a JNC install ISO for Jason-Desktop-Linux was about 100GiB,
so it is definitely possible in a worst case scenario.

**I’m just guessing here. I have no idea how much it would actually be.

[1]: <https://github.com/NixOS/nixpkgs/blob/e402c3eb6d88384ca6c52ef1c53e61bdc9b84ddd/nixos/modules/installer/cd-dvd/iso-image.nix#L508-L521>
[2]: <https://github.com/NixOS/nixpkgs/blob/e402c3eb6d88384ca6c52ef1c53e61bdc9b84ddd/nixos/modules/installer/cd-dvd/iso-image.nix#L814-L818>
[3]: <https://github.com/NixOS/nixpkgs/blob/e402c3eb6d88384ca6c52ef1c53e61bdc9b84ddd/nixos/modules/installer/cd-dvd/iso-image.nix#L830>
[4]: 88f3d89 (Disable squashfs compression on installation ISOs, 2024-02-28)