Install smem on all systems
There’s a bcachefs bug that causes the kernel to use much more memory
than it should [1]. I think that the bug has been fixed in the latest
unstable version of Linux, but not in the latest stable version of
Linux.

I often use “smem -twk” to check how much memory my system is using. If
smem tells me that the kernel is using a lot of RAM, then I use the
workaround from this GitHub issue [1] to make the kernel use less.

In order to run smem, I run “nix-shell -p smem”. This commit makes it so
that I don’t have to run nix-shell in order to run smem.

[1]: <https://github.com/koverstreet/bcachefs/issues/384>