Revert "Enable nix-ld"
This reverts commit 6cf759e25b744ba4c1f47e8027bb75057afe3e7a.

The internal documentation site has been migrated to Docusaurus. Now
that that site no longer uses Ruby, there’s no longer a good reason to
have /lib/ld-linux.so.2 on my system.

Additionally, having that file on my system means that I run the risk of
accidentally depending on it for other things. For example, I noticed
recently that the pre-commit tool was downloading a copy of go and
running it. I don’t want pre-commit to do that. I want pre-commit to
just give an error if something isn’t specified declaratively in my
NixOS config. I want everything to be as Nix-pure as possible.