Restart from a blank slate
If Jason’s NixOS config was first created today, then it would look a
lot different. According to this repo’s commit history, I first created
my NixOS config before 2022-06-09 [1]. I’ve learned a lot since then.

One of the things that I’ve learned about recently is flakes. There’s a
lot of different advantages to using flakes, but I’m mainly concerned
about reproducibility. Using flakes forces me to pin the version of
Nixpkgs that I’m using. If I ever have to use git-bisect on this repo,
I’ll be able to more accurately judge which commit introduced a
particular regression. Also, I’ll be able to hold off on a Nixpkgs
version bump if a Nixpkgs commit breaks my config somehow.

The main motivation behind this change is to make it easier to start
using flakes for my NixOS config, but I’m sure that I’ll make a lot of
improvements as a side effect of this change.

This commit message is dedicated to the public domain using 🅭🄍1.0 [2].

[1]: 1a76879 (Import config from old repo, 2022-06-09)
[2]: <https://creativecommons.org/publicdomain/zero/1.0>