Create copying.md
I have a specific style that I want all of my repos to use. I created a
tool called repo-style-checker [1] that checks to see if a repo follows that
style.

Before this commit, if I ran repo-style-checker, then it would give me
an error that says that this repo doesn’t have a copying.md file. This
commit fixes that error.

This commit message is dedicated to the public domain using 🅭🄍1.0 [2].
The copying.md file added by this commit is dedicated to the public
domain using 🅭🄍1.0 [2].

[1]: <https://github.com/Jayman2000/jasons-pre-commit-hooks/blob/dc5604486148f209f06b387a8caf545d80627384/pyproject.toml#L41>
[2]: <https://creativecommons.org/publicdomain/zero/1.0>