Enable repo-style-checker pre-commit hook
I have a specific style that I want all of my repos to use. I created a
tool called repo-style-checker [1] that checks to see if a repo follows
that style.

Before this commit, I had to remember to run repo-style-checker
manually. This commit makes it harder for me to forget to run
repo-style-checker.

[1]: <https://github.com/Jayman2000/jasons-pre-commit-hooks/blob/dc5604486148f209f06b387a8caf545d80627384/pyproject.toml#L41>