Explain how to set up pre-commit
Before this change, if you ran “pre-commit run --all”, it would fail.
Specifically, the repo-style-checker hook would give many errors. Two of
those errors had to do with the fact that the README didn’t explain how
to set up pre-commit. This commit fixes those errors.

Additionally, it’s important to explain to potential contributors how to
contribute to a project and how to do a good job at contributing to a
project.