Create dev shell for working on this repo
This dev shell will make it easier for me to work on this repo. It will
allow me to start working on the repo without having to think about
whether or not I have the right stuff installed my system. All I’ll need
in order to start working on this repo is Nix.