Rename pkgs to pkgsForThisFlake
The goal of this commit is to eventually make this repo easier to
understand. A future commit is going to make this repo’s flake output a
NixOS config. When I create that NixOS config, I’m probably going to
have a different attribute named pkgs. I say that because pkgs is in
included in the attribute set that gets passed to NixOS modules that are
functions [1].

[1]: <https://nixos.org/manual/nixos/stable/#sec-writing-modules>