Move common options into separate module
The idea is that I’ll eventually create an option named jnc.machineSlug.
The only thing that the files in modules/configuration.nix/ will have to
do is set jnc.machineSlug. Everything else will be handled by the
modules in modules/imports/. Hopefully, doing that will make the config
easier to work with.