Pin Nix version
The NixOS manual says [1]:

> Warning
> It is generally safe to switch back and forth between channels. The
> only exception is that a newer NixOS may also have a newer Nix
> version, which may involve an upgrade of Nix’s database schema. This
> cannot be undone easily, so in that case you will not be able to go
> back to your original channel.

In future commits, I’m going to try to upgrade systems to NixOS 24.11.
In the past, upgrading to a new NixOS version has broken things for me,
and I’ve had to temporarily go back to previous versions of NixOS in
order to fix things. If one of those database upgrades were to happen,
then I wouldn’t be able to go back to a previous NixOS version.

I’m pining the Nix version now in order to ensure that a database
upgrade doesn’t happen when I upgrade to NixOS 24.11. That way, I’ll be
sure that I can go back to NixOS 24.05 if things break.

[1]: <https://nixos.org/manual/nixos/stable/#sec-upgrading>