Remove sound.enable
The main motivation behind this change is to make it easier to create a
future commit. That future commit will upgrade Graphical-Test-VM to
NixOS 24.11. If you try to set sound.enable in a NixOS 24.11 config,
then you’ll get this error:

> Failed assertions:
> - The option definition `sound' in
> `/nix/store/[…]-jason-nixos-config/modules/graphical.nix' no longer
> has any effect; please remove it.
> The option was heavily overloaded and can be removed from most
> configurations.

Removing sound.enable now means that we won’t have to remove it later.