Remove hardware.pulseaudio.enable
The main motivation behind this change is to make it easier to create a
future commit. That future commit will upgrade Graphical-Test-VM to
NixOS 24.11. Having hardware.pulseaudio.enable set to true was causing
this error on NixOS 24.11:

> Failed assertions:
> - Using PipeWire as the sound server conflicts with PulseAudio. This
> option requires `hardware.pulseaudio.enable` to be set to false

Removing hardware.pulseaudio.enable now means that we won’t encounter
that error later.