Update Disko
The main motivation behind this change is to make it easier to create a
future commit. That future commit doesn’t make any changes that are
directly related to Disko, disk partitioning or filesystems, but the
changes were causing automated installations to fail. Specifically, the
changes in that future commit caused disko-install to fail to set a
partitions name. I don’t know why that future commit is failing.
Hopefully, updating Disko will cause that future commit to start working
again.