Fix incorrect escaping in randfd
Before this change, the final command in randfd’s pipeline did the wrong
thing. It was supposed to replace NUL characters with LF characters. In
reality, it was replacing backslash characters with backslash characters
(which is redundant) and replacing 0 characters with lowercase n
characters. This change fixes that problem so that randfd works as
intended.