Lock major and minor Nix version
NixOS 24.11 is now deprecated [1]. At the moment, all of the NixOS
configs in this repository use NixOS 24.11. Those configs should
eventually be updated so that they don’t use a deprecated version of
NixOS.

Here’s a quote from the NixOS 25.05 manual’s “Upgrading NixOS” section
[2]:

> It is generally safe to switch back and forth between channels. The
> only exception is that a newer NixOS may also have a newer Nix
> version, which may involve an upgrade of Nix’s database schema. This
> cannot be undone easily, so in that case you will not be able to go
> back to your original channel.

This change makes sure that all of the configs in this repository use
Nix version 2.24.x regardless of what version of NixOS they use. This
will help ensure that I don’t run into problems if I need to switch back
and forth between versions of NixOS. While working on upgrading to NixOS
25.05, I likely will need switch back and forth between versions of
NixOS as I discover and fix problems.

Once all of the configs in this repository have been upgraded to use
NixOS 25.05, this commit will no longer be necessary or helpful. After
that happens, this commit should be reverted.

[1]: <https://nixos.org/blog/announcements/2025/nixos-2505>
[2]: <https://nixos.org/manual/nixos/stable/#sec-upgrading>