Lock Linux kernel version
Before this change, the configs in this repository used
linuxPackages_latest. At the moment, linuxPackages_latest provides Linux
kernel version 6.14.8. Soon, I’m going to try to upgrade from NixOS
24.11 to NixOS 25.05. In NixOS 25.05, linuxPackages_latest provides a
newer version of the Linux kernel. Sometimes, switching between versions
of the Linux kernel can take a very long time on my systems due to
bcachefs-related changes. Sometimes, I end up having to boot into a USB
drive and manually doing an fsck. I don’t want to have to do that when
switching between versions of NixOS so I’m locking the version now. I’ll
unlock the version after I’ve finished upgrading all of the configs in
this repository to NixOS 25.05.

I chose to use linuxPackages_6_12 because on NixOS 25.05 the default
version of the Linux kernel is 6.12 at the moment. This means that I’ll
be able to completely get rid of the boot.kernelPackages part of the
configs after I’ve finished upgrading to NixOS 25.05.