Use kalendar package’s new name
Ever since this pull request [1] was merged into Nixpkgs, the kalendar
attribute has just been an alias for the merkuro package. This change
makes it so that Jason’s NixOS Config uses the main merkuro attribute
instead of the kalendar alias

The main motivation behind this change is to make it easier to make a
future commit. While working on that future commit, I was getting an
error about merkuro, and I wasn’t sure why because this repository
didn’t contain any references to merkuro. Using the merkuro attribute
directly instead of using the kalendar alias will make that error easier
to understand.

[1]: <https://github.com/NixOS/nixpkgs/pull/251163>