Ensure flake’s store path is present on install mediums
Before this change, if you tried to perform an unattended install using
the code that’s in this repository, then you would get an error telling
you that a certain Nix store path does not exist. You can find out more
about that error here [1].

This change fixes that error by (hopefully) ensuring that the required
Nix store path is guaranteed to be in install mediums’s Nix stores.

[1]: <https://discourse.nixos.org/t/am-i-understanding-string-contexts-correctly/65567?u=jasonyundt>