Make copy and pasting work in Neovim
There are a few different packages that I could have installed in order
to make copy and pasting work. I chose wl-clipboard-rs because it’s
native to Wayland and it’s written in Rust. The other Wayland copy and
paste tools that I saw were written in the C Programming Language. If I
ever run into a problem, then it’s going to be much easier for me to
contribute to a project written in Rust than it would be for me to
contribute to a project written in C.