Fix potential Syncthing file permission errors
Before this change, users in the syncthing group could create files in
synced directories that were not owned by the syncthing user or the
syncthing group. If you create a file in a synced directory that is not
owned by the syncthing user or the syncthing group, then Syncthing
itself won’t be able to sync that file.

This change pretty much forces all of the files in synced directories to
be owned by the syncthing group. If all of the files are owned by the
syncthing group, then the previously described issue cannot happen.