Increase services.syncthing.settings.options.urAccepted
Before this change, whenever you set up a new system with a NixOS
configuration from this repository, Syncthing would ask you to confirm
using a new format for anonymous usage reporting. This change makes it
so that you don’t have to bother confirming the new format. The new
format will be confirmed automatically for you.