Allow installer to boot via USB
Before this change, when you created an installation medium for
graphical-test-vm, you had to connect the installation medium to the
virtual machine as a VirtIO device. If you connect it as a virtual USB
device, then it will fail to boot. This change fixes that issue.

For virtual machines, it’s OK if the installation medium must be
connected via VirtIO. However, I eventually want to create a config in
this repository that will get installed on a physical machine. I need
USB boot to work or else I won’t be able to use the installation medium
on physical hardware.