Make run0 remember password temporarily
This is supposed to make it so that you can run a few different commands
via run0 without it asking for your password each time. Instead of
asking for your password each time, run0 will ask you for your password
once and then remember it for a bit.

Unfortunately, this change doesn’t work at the moment. It should start
working after we switch to a newer version of polkit (specifically, we
need a version of polkit that has this pull request [1] merged into it).

[1]: <https://github.com/polkit-org/polkit/pull/533>