Enable nix-check-flake pre-commit hook
The main motivation behind this change is to prevent a future error. In
a future commit, I’m going to update the version of
jasons-pre-commit-hooks that this repository uses. Specifically, I’m
going to update jasons-pre-commit-hooks to version 0.5.0. Here’s a quote
from the patch notes for version 0.5.0 of jasons-pre-commit-hooks [1]:

> • repo-style-checker now requires that you use the nix-check-flake and
> the nix-format pre-commit hooks.

Enabling the nix-check-flake pre-commit hook now means that I won’t get
errors about it after I update jasons-pre-commit-hooks to version 0.5.0.

[1]: <https://github.com/Jayman2000/jasons-pre-commit-hooks/blob/v0.5.0/CHANGELOG.md#added>