Upgrade to Linux 6.15
The main motivation behind this change is to make it easier to create a
future commit. The future commit is going to update flake.lock so that
we use a newer commit from Nixpkgs’s nixos-25.05 branch. Linux 6.14 has
been removed from Nixpkgs’s nixos-25.05 branch [1]. Upgrading to a newer
version of Linux now means that we won’t have to upgrade to a newer
version of Linux later.

[1]: <https://github.com/NixOS/nixpkgs/pull/420276#event-18796022561>