Set system.configurationRevision
This will allow me to use “nixos-rebuild list-generations in order to
determine which version of this repository a particular machine is
using. Being able to easily figure out which version of this repository
is being used will allow me to make sure that I didn’t forget to deploy
the newest version of this repository on a particular machine.