Move configuration of Nix into separate module
Before this change, the configuration for the Nix package manager was
put in this repository’s flake’s default NixOS module. This change moves
the configuration for the Nix package manager into a separate module and
makes it so that the default module imports this new module.

The main motivation behind this change is to keep things organized in
the long term. In future commits, I plan to make additions to Nix’s
configuration. Putting Nix’s configuration in its own module makes it so
that there isn’t so much random stuff in the default module. The default
module already contains too much random stuff.