Rename ssh-server NixOS module to ssh
That module doesn’t just configure an SSH server. It also configures an
SSH client. Calling it just “ssh” instead of “ssh-server” is more
accurate.

The main motivation behind this change is to make it easier to create a
future commit. That future commit will add additional SSH client
configuration to this module. Changing the name of the module now means
that the module’s name won’t become even more inappropriate in the
future.