Stop pinning the Linux kernel version
The version of NixOS’s bcachefs module that’s in newer revisions of
Nixpkgs’s nixos-25.05 branch will automatically set the kernel version
for you [1]. It will probably be better to follow the version that the
module uses than to use our own version.

Additionally, the version of NixOS’s bcachefs module that’s in newer
revisions of Nixpkgs’s nixos-25.05 branch use the bcachefs DKMS module
rather than the built-in kernel module [2]. This means that pinning the
kernel version will not prevent bcachefs upgrades. The only reason why I
was pinning the kernel version was to delay bcachefs upgrades.

This change is being made in order to prepare for a future commit. In
that future commit, we will start using a newer revision of Nixpkgs’s
nixos-25.05 branch.

[1]: <https://github.com/NixOS/nixpkgs/blob/09d3831f44af7689a2bc1b190ee69160561da363/nixos/modules/tasks/filesystems/bcachefs.nix#L246>
[2]: <https://discourse.nixos.org/t/bcachefs-with-future-kernel-versions/71713>