Enable Docker
I run Syncthing on a device that doesn’t use the Nix package manager at
all. I use Docker in order to run Syncthing on that device. That device
is not able to download Docker images on its own. As a result, I have to
run “docker pull” on one machine and then manually copy the image to the
other device whenever I want to update Syncthing on that device. I’m
enabling Docker so that I can do that.